Removed return of image objects from functions to reference sooner
When images are passed back as return objects from functions, their reference stays alive until they are deferenced.  It is cleaner to dereference them as soon as they are no longer needed or let them get cleaned up when a function is exited.  This is tied to the issue of raster files or rdc files not getting fully written until they are dereferenced.