Implemented New Plotting Style and Added MedAE Percentage Calculation
This update introduces a new plotting style that incorporates a legend and a white boundary around each dot to improve visibility of overlaps, enhancing the visual clarity of our plots. Additionally, we've integrated the calculation of the MedAE percentage. This is calculated as (MedAE/Grid_Size) * 100, where the Grid_Size represents the area of the assessment grids in hectares (ha), with a default value of 100,000 ha. The resulting MedAE% is expected to range between 0.1% and 0.9%.