Added (UDef-ARP.py) binary map checker function
Added a binary map checker to automatically check whether the input data 
is a binary map or not. After the function returns False, an error 
message will show up.

A binary map (boolean map) is characterized by a data type of either 
byte or integer, with minimum and maximum values set to 0 and 1. The 
only exception is when the data type is real, with minimum and maximum 
values precisely at 0.0000000 and 1.0000000 and would need to check 
pixel values.

The validation logic is as follows:

1.	If data type is byte or integer and min is 0 and max is 1, it returns 
True.

2.	If data type is real and min is 0 and max is 1, it executes a 
function to check the number of unique values. Because the function only 
run under this condition and break when the unique value larger than 2, 
it executed fast in my testing.
   (1)	If the unique number is 2 [0,1], return True.
   (2)	If the number is larger than 2, break and return False.

3.	For any other scenario, it returns False.