Fixed Adjusted Relative Frequency Table Issue in allocation_tool.py
In cases where bins are missing during the prediction phase, the original adjusted relative frequency table is merge CAL relative frequency table and the missing bins table. However, some bins in the CAL table do not exist in the CNF modeling region, but this does not impact the final prediction since the adjusted table is only used for assigning values.
This update cleans up bins in the Adjusted Relative Frequency Table that are not present in the CNF modeling region.